+++ src/components/Elements/FloatingContent/FloatingContent.spec.js
@@ -25,7 +25,7 @@ describe('FloatingContent component', () => {
 
   it('should have default classes', () => {
     render(
-      <FloatingContent onToggle={noop} trigger={Trigger} isShown className="foo">
+      <FloatingContent onToggle={noop} trigger={Trigger} isShown className="foo" reversePlacementOnSmallSpace={false}>
         some text here
       </FloatingContent>,
     );
@@ -89,7 +89,13 @@ describe('FloatingContent component', () => {
 
   it('should allow to define the floating content placement', () => {
     const { rerender } = render(
-      <FloatingContent onToggle={noop} trigger={Trigger} isShown placement="top-left">
+      <FloatingContent
+        onToggle={noop}
+        trigger={Trigger}
+        isShown
+        placement="top-left"
+        reversePlacementOnSmallSpace={false}
+      >
         some text here
       </FloatingContent>,
     );
@@ -100,7 +106,13 @@ describe('FloatingContent component', () => {
     expect(floatingContent.classList.contains('float-top-left')).to.be.true;
 
     rerender(
-      <FloatingContent onToggle={noop} trigger={Trigger} isShown placement="bottom-right">
+      <FloatingContent
+        onToggle={noop}
+        trigger={Trigger}
+        isShown
+        placement="bottom-right"
+        reversePlacementOnSmallSpace={false}
+      >
         some text here
       </FloatingContent>,
     );
@@ -109,7 +121,7 @@ describe('FloatingContent component', () => {
     expect(floatingContent.classList.contains('float-bottom-right')).to.be.true;
 
     rerender(
-      <FloatingContent onToggle={noop} trigger={Trigger} isShown>
+      <FloatingContent onToggle={noop} trigger={Trigger} isShown reversePlacementOnSmallSpace={false}>
         some text here
       </FloatingContent>,
     );
@@ -129,5 +141,20 @@ describe('FloatingContent component', () => {
     const floatingContent = document.querySelector('#bi-floats .bi.bi-floater');
 
     expect(floatingContent.getAttribute('style')).not.equal('margin: 10px;');
+  });
+
+
+  it('should change placement if there\'s no enough space to show the floating content', () => {
+    render(
+      <FloatingContent onToggle={noop} trigger={Trigger} isShown className="foo" reversePlacementOnSmallSpace>
+        some text here
+      </FloatingContent>,
+    );
+
+    const floatingContent = document.querySelector('#bi-floats > div');
+
+    expect(floatingContent.classList.contains('bi')).to.be.true;
+    expect(floatingContent.classList.contains('bi-floater')).to.be.true;
+    expect(floatingContent.classList.contains('float-bottom-center')).to.be.true;
   });
 });